🔨 Completed Part4

+++ Part4/blog-list/controllers/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
 const jwt = require("jsonwebtoken");
 const config = require("../utils/config");
+const mongoose = require("mongoose");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", {
@@ -10,22 +11,30 @@ blogsRouter.get("/", async (request, response) => {
     name: 1,
     id: 1,
   });
+
   if (blogs) {
     response.json(blogs);
   } else {
-    response.status(404).end();
+    response.status(400).end();
   }
 });
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
   const token = request.token;
-  const decodedToken = jwt.verify(token, config.SECRET);
+
+  let decodedToken;
+  try {
+    decodedToken = jwt.verify(token, config.SECRET);
+  } catch (error) {
+    return response.status(401).json({ error: "Invalid Token" });
+  }
+
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" });
   }
 
-  const user = await User.findById(body.userId);
+  const user = request.user;
 
   if (!body.title || !body.url) {
     return response.status(400).json({
@@ -57,8 +66,11 @@ blogsRouter.delete("/:id", async (request, response) => {
 
   const id = request.params.id;
   const blog = await Blog.findById(id);
-  const user = await User.findById(decodedToken.id);
-  console.log(user);
+  const user = request.user;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return response.status(400).json({ error: "invalid ID" });
+  }
 
   if (blog.user.toString() === user.id.toString()) {
     await Blog.deleteOne({ _id: id });
@@ -69,6 +81,10 @@ blogsRouter.delete("/:id", async (request, response) => {
 });
 
 blogsRouter.get("/:id", async (request, response) => {
+  const id = request.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return response.status(400).json({ error: "invalid ID" });
+  }
   const blog = await Blog.findById(request.params.id);
 
   if (blog) {
@@ -79,6 +95,10 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.put("/:id", async (request, response) => {
+  if (!mongoose.Types.ObjectId.isValid(request.params.id)) {
+    return response.status(400).json({ error: "invalid ID" });
+  }
+
   const { likes } = request.body;
   console.log({ likes });
   const updateBlog = await Blog.findByIdAndUpdate(
